L2-992: Add testing
