+++ frontend/src/lib/components/accounts/AccountsTitle.svelte
@@ -19,7 +19,7 @@
 </script>
 
 <div class="title">
-  <h1>{$i18n.accounts.title}</h1>
+  <h1 data-tid="accounts-title">{$i18n.accounts.title}</h1>
 
   {#if balance !== undefined}
     <Tooltip
